2.0.0
Major update.

All existing functionality is part of a (now renamed) "SimpleMiniMvcs" base class that will remain supported. There are 'beginner' samples available within the package manager for this.

All new functionality is part of a (new) "MiniMvcs" base class that is more robust. There are 'advanced' samples available within the package manager for this.